Rewrite buffering and argument parsing
This gives the user the ability to pass command line arguments to
asha_pipewire_sink.

One of the arguments is the buffering algorithm, of which I have
implemented five variants. Without the ability to monitor the l2cap
credit depth, we have to use other means to control the stream
stability, and they will vary depending on the hearing device.